v5修改

+++ yudao-module-steam/yudao-module-steam-biz/src/main/java/cn/iocoder/yudao/module/steam/service/fin/v5/vo/V5BuyProductVo.java
@@ -8,13 +8,13 @@ import java.math.BigDecimal;
 @Data
 public class V5BuyProductVo implements Serializable {
     private String marketHashName;
-    private Double purchasePrice;
+    private BigDecimal purchasePrice;
     private String tradeUrl;
     private String merchantOrderNo;
     private String merchantKey;
     private int buyType;
 
-    public V5BuyProductVo(String marketHashName, Double purchasePrice, String tradeUrl, String merchantOrderNo, String merchantKey, int buyType) {
+    public V5BuyProductVo(String marketHashName, BigDecimal purchasePrice, String tradeUrl, String merchantOrderNo, String merchantKey, int buyType) {
         this.marketHashName = marketHashName;
         this.purchasePrice = purchasePrice;
         this.tradeUrl = tradeUrl;
